implement strategy

+++ rust_strategy/src/params.rs
@@ -1,10 +1,10 @@
 #[derive(Clone, Debug)]
 pub struct Params {
     pub lookback_zscore: usize,
-    pub lookback_performance: usize,
-    pub entry_z: f64,
-    pub exit_z: f64,
-    pub expiry_close_days: usize, // days before contract's last appearance to force close
+    pub lookback_performance: u32,
+    pub entry_z: f32,
+    pub exit_z: f32,
+    pub expiry_close_days: u32, // days before contract's last appearance to force close
     pub debug: bool,
     pub commodity_a_prefix: String,
     pub commodity_b_prefix: String,
